# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GOERS",
        "description": "REST Api Documentation - Representational State Transfer Protocol",
        "contact": {
            "name": "GOERS",
            "email": "contact@goers.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/v1/organization": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get item of organization",
                "description": "",
                "operationId": "getOrganization",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/RequestedWith"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Successful"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/GeneralError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Create new organization",
                "description": "",
                "operationId": "storeOrganization",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/RequestedWith"
                    }
                ],
                "requestBody": {
                    "description": "Organization",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrganizationPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Successful"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/GeneralError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "OrganizationPostRequest": {
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Successful": {
                "description": "Successful Operation"
            },
            "Unauthorized": {
                "description": "Unauthorized"
            },
            "Forbidden": {
                "description": "Forbidden"
            },
...
            "BadRequest": {
                "description": "Bad Request"
            },
...
            "GeneralError": {
                "description": "General Error"
            }
        },
        "parameters": {
            "RequestedWith": {
                "name": "X-Requested-With",
                "in": "header",
                "description": "Determinate as ajax request",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "XMLHttpRequest",
                    "enum": [
                        "XMLHttpRequest"
                    ]
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}